requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
pillow
opencage
python-decouple
channels
channels-redis
daphne
redis
requests
gunicorn